Api layer view structures is changed.

+++ Cybersec.Service/ViewModels/Article/ContentBlockViewModel.cs
@@ -1,6 +1,6 @@
 ﻿namespace Cybersec.Service.ViewModels.Article;
 public class ContentBlockViewModel
 {
-    public int Id { get; set; }
-    public int Order { get; set; }
+    public long Id { get; set; }
+    public long Order { get; set; }
 }
